+++ lib/td_se_web/endpoint.ex
@@ -5,7 +5,7 @@ defmodule TdSeWeb.Endpoint do
   #
   # You should set gzip to true if you are running phoenix.digest
   # when deploying your static files in production.
-  plug(Plug.Static, at: "/", from: :td_se, gzip: false, only: ~w(swagger.json))
+  plug(Plug.Static, at: "/", from: :td_se, gzip: false)
 
   # Code reloading can be explicitly enabled under the
   # :code_reloader configuration of your endpoint.
@@ -34,19 +34,4 @@ defmodule TdSeWeb.Endpoint do
   )
 
   plug(TdSeWeb.Router)
-
-  @doc """
-  Callback invoked for dynamically configuring the endpoint.
-
-  It receives the endpoint configuration and checks if
-  configuration should be loaded from the system environment.
-  """
-  def init(_key, config) do
-    if config[:load_from_system_env] do
-      port = System.get_env("PORT") || raise "expected the PORT environment variable to be set"
-      {:ok, Keyword.put(config, :http, [:inet6, port: port])}
-    else
-      {:ok, config}
-    end
-  end
 end